Update physical exposure and social fabric maps
Public sent an email notifying us of discrepancies between maps on OpenDRR
GitHub Pages and data on FGP.  Upon further discovery, the maps were linking
to an older version of ESRI REST layer and not the newest version.

Special thanks to the member of the general public for the report,
and to William Chow (@wkhchow) for the investigation and diagnostics.

Fixes OpenDRR/opendrr.github.io#32, OpenDRR/national-human-settlement#40